+++ NEWS.md
@@ -1,5 +1,7 @@
 # pkgdown (development version)
 
+* Setting `clean = TRUE` in `deploy_site_github()` removes old files from the deployed site before building a new one (#1297).
+
 # pkgdown 1.5.1
 
 * Syntax highlighting works on Windows once more (#1282).